requirements.txt
numpy>=1.24
scikit-learn>=1.3
librosa>=0.10
torch>=2.0
torchaudio>=2.0
speechbrain>=0.5.16
faiss-cpu>=1.7.4
joblib>=1.3
yt-dlp>=2024.3.10

